個別Amazon商品リンクキャッシュの削除機能追加

+++ lib/page-cache/cache-func.php
@@ -10,10 +10,7 @@
 if ( !function_exists( 'delete_sns_count_caches' ) ):
 function delete_sns_count_caches(){
   if (is_user_administrator()) {
-    $res =  delete_db_cache_records(TRANSIENT_SHARE_PREFIX);
-    if ($res) {
-      return delete_db_cache_records(TRANSIENT_FOLLOW_PREFIX);
-    }
+    return delete_db_cache_records(TRANSIENT_SHARE_PREFIX) && delete_db_cache_records(TRANSIENT_FOLLOW_PREFIX);
   }
 }
 endif;
@@ -48,9 +45,50 @@ endif;
 //全てのキャッシュを削除
 if ( !function_exists( 'delete_all_theme_caches' ) ):
 function delete_all_theme_caches(){
-  delete_sns_count_caches();
-  delete_popular_entries_caches();
-  delete_blogcard_caches();
-  delete_amazon_api_caches();
+  return delete_sns_count_caches() &&
+         delete_popular_entries_caches() &&
+         delete_blogcard_caches() &&
+         delete_amazon_api_caches();
+}
+endif;
+
+//Amazon個別キャシュの削除
+if ( !function_exists( 'delete_amazon_asin_cache' ) ):
+function delete_amazon_asin_cache($asin){
+  if (is_user_administrator()) {
+    return delete_db_cache_records(TRANSIENT_AMAZON_API_PREFIX.$asin) &&
+           delete_db_cache_records(TRANSIENT_BACKUP_AMAZON_API_PREFIX.$asin);
+  }
+}
+endif;
+
+//テーマが保管している全てのキャッシュの保存処理
+if ( !function_exists( 'delete_theme_storaged_caches' ) ):
+function delete_theme_storaged_caches(){
+  //管理者権限を持っているログインユーザーかどうか
+  if (is_user_administrator()) {
+    $delete_option = isset($_GET['cache']) ? $_GET['cache'] : null;
+    switch ($delete_option) {
+      case 'all_theme_caches':
+        return delete_all_theme_caches();
+        break;
+      case 'sns_count_caches':
+        return delete_sns_count_caches();
+        break;
+      case 'popular_entries_caches':
+        return delete_popular_entries_caches();
+        break;
+      case 'blogcard_caches':
+        return delete_blogcard_caches();
+        break;
+      case 'amazon_api_caches':
+        return delete_amazon_api_caches();
+        break;
+      case 'amazon_asin_cache':
+        $asin = isset($_GET['asin']) ? $_GET['asin'] : null;
+        return delete_amazon_asin_cache($asin);
+        break;
+    }
+  }
 }
 endif;
